requirements.txt
# GNN_PPO_VEC Dependencies
# 图神经网络增强VEC任务调度系统依赖

# 深度学习框架
torch>=1.13.0
torch-geometric>=2.3.0

# 数值计算
numpy>=1.21.0
scipy>=1.9.0

# 数据处理
pandas>=1.5.0
scikit-learn>=1.1.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.11.0
networkx>=2.8.0

# 配置管理
PyYAML>=6.0

# 实用工具
tqdm>=4.64.0
argparse

# 可选依赖（用于更高级的可视化）
plotly>=5.0.0
dash>=2.0.0

# 开发工具（可选）
jupyter>=1.0.0
ipykernel>=6.0.0

# 测试工具（可选）  
pytest>=7.0.0
pytest-cov>=4.0.0